Specialize ActivityListener for ClientConnection
Only `ClientConnection` is able to notify `ActivityListener` since only `ClientConnection` is responsible for reading
and writing messages. The registration methods have been moved from `Connection` to `ClientConnection`, with the
`notifyListeners()` helper method being also moved from `AbstractConnection` to `AbstractClientConnection`.
